web-animations/animation/animation-constructor-target.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation constructor target argument test</title>
<meta name="assert" content="The animation target or target pseudo-element. This may be null for animations that do not target a specific element">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-animation-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var animation = new Animation(null, null);

    assert_equals(animation.target, null, 'Animation.target should be null');
}, 'Animation constructor called with null value of target argument creates ' +
    'Animation object instance with null target');

test(function() {
    var target = createDiv(this);
    var animation = new Animation(target, null);

    assert_equals(animation.target, target, 'Animation.target should be set by ' +
        'the constructor');
}, 'Animation constructor called with not null value of target argument creates ' +
    'Animation object instance with specified target');

test(function() {
    var element1 = createDiv(this);
    var element2 = createDiv(this);
    var pseudoElement = element2.pseudo(':before');

    var animation = new Animation(pseudoElement, null);

    assert_equals(animation.target, pseudoElement, 'Animation.target should be set ' +
        'by the constructor');
}, 'Animation constructor called with pseudo-element as value of target argument creates ' +
    'Animation object instance with specified PseudoElement as a target');
</script>
</body>
